+++ saleor/payment/interface.py
@@ -324,11 +324,6 @@ class GatewayResponse:
     transaction_already_processed: bool = False
     psp_reference: str | None = None
 
-    # Temporary pass Adyen-plugin-specific data to model, so we can drop raw_response
-    # After the plugin is gone, this should be removed
-    legacy_adyen_plugin_result_code: str | None = None
-    legacy_adyen_plugin_payment_method: str | None = None
-
 
 @dataclass
 class AddressData: