+++ saleor/payment/migrations/0067_remove_transaction_legacy_adyen_plugin_fields.py
@@ -1,0 +1,18 @@
+from django.db import migrations
+
+
+class Migration(migrations.Migration):
+    dependencies = [
+        ("payment", "0066_transactionitem_gift_card"),
+    ]
+
+    operations = [
+        migrations.RemoveField(
+            model_name="transaction",
+            name="legacy_adyen_plugin_result_code",
+        ),
+        migrations.RemoveField(
+            model_name="transaction",
+            name="legacy_adyen_plugin_payment_method",
+        ),
+    ]